1.1.14 + 26 사진 추천 api 추가완료, 마이페이지 오버플로우 해결, read 페이지 멘트 추가, 알림페이지 숨김, 알림 gif 자르기

+++ lib/features/home/controller/bukkung_list_page_controller.dart
@@ -98,9 +98,9 @@ class BukkungListPageController extends GetxController {
 
   void deleteBukkungList(
       BukkungListModel bukkungListModel, bool isDeleteImage) async {
-    final GroupModel groupModel = AuthController.to.group.value;
     if (isDeleteImage) {
-      if (Constants.baseImageUrl != bukkungListModel.imgUrl) {
+      if (Constants.baseImageUrl != bukkungListModel.imgUrl &&
+          bukkungListModel.imgUrl!.startsWith('https://firebasestorage')) {
         await BukkungListRepository()
             .deleteListImage('${bukkungListModel.imgId}.jpg');
       }
